Correctly remove new empty message subtition
I still dont know clearly how the database contexes work. Local stores value that will be staged to database context once save is called. This check previously just checked if we have already saved empty field and not if we are going to save one. Also empty pattern is null and not "" empty string.